move llm.thread model's codes
