+++ llm_thread/models/__init__.py
@@ -1,3 +1,2 @@
 from . import llm_thread
 from . import mail_message
-from . import llm_thread_utils
