+++ llm_thread/models/mail_message.py
@@ -1,5 +1,6 @@
 import json
 import logging
+
 import markdown2
 
 from odoo import _, api, fields, models
